Рефакторинг класса ApiSecurityTests на GetApiSecurityTestsand PostApiSecurityTests, Рефакторинг Пэкеджа utilsRefactoring: Split ApiSecurityTests into GetApiSecurityTests and PostApiSecurityTests, restructure utils package
